mnesia_db: continue development of rest-api and database
The mnesia_db has been improved. The asn1_handler can now fetch work
items from the database and store states in there. At the moment the
BaseUrl is the only state we store in the database. The basic
functionaliy is now there, however there is still some room for
improvelemts. In particular we need a way to store some diagnostic info
we can expose on the rest info. This is in theory already possible,
since when we finish work we can put an "outcome" that is returned to
the REST api. However this outcome is more intended to return structured
data like the results from PSMOs etc.

The database initialization has been improved as well. The process is
now fully automatic, so there is no need to call mnesia_db:setup() on
startup.