esipa_asn1_handler: highlight places where we know the TransactionId
We currently use the PID of the HTTP connection to determine to which
transaction/work an HTTP request belongs. This method is only reliable
in situations where the HTTP connection is unterrupted. Unfortunately we
can not use the TransactionId as identifier alone since it is not
contained in all transactions, However, we can use the TransactionId in
addition to the PID to increase the robustness of the communication.

As a preperation, let's highlight all places, where we know the
TransactionId