esipa_asn1_handler: improve clarity of JSON REST API
The json API has PSMO objects that are organized in a list. The objects
themselves have no field names in front (there is a name field inside).
This makes the PSMO type dificult to identify before looking into the
PSMO object. This makes it difficult to describe the JSON structure in
an abstract language.

There is also another peculiarity: depending on the facility
we find a download_order or a psmo_order field. When describing this in
an abstract way it is difficult to express that there can only be either
a download_order or a psmo_order in the JSON structure.

This patch fixes those problems and leaves us with a much clearer JSON
syntax that is much easier to describe.

Related: SYS#6720