contrib/tryme_add/deleteEim: use separate eimId for testing
The eimConfiguration and we use for testing uses the same eimId we
use for the regular test operation. This is too interfering. Let's
use a separate eimId. Then we can test the adding and deleting easily
on the side without disturbing our useful eimConfiguration.

(the changed eimConfiguration we use here also corrects the wrongly
applied public key, but this has no effect here)

Related: SYS#6720