mnesia_db: fix memory/database leak
When deleting an order via the REST API we delete the related item from
the rest table, but we do not check if there is a coresponding work item
in the work table. When this is the case, the related work item would
become orphaned and would stay in the work table indefinetly. We must
make sure that we never leave an orphaned item in the work table.

Related: SYS#6720